+++ test/linker.test.ts
@@ -1,7 +1,7 @@
 import { expect, should, use } from 'chai'
-import { Body, Class, Closure, Describe, Environment, Field, Import, Method, Mixin, NamedArgument, Package, Parameter, ParameterizedType, Reference, Return, Singleton, Test, Variable } from '../src/model'
 import { fromJSON } from '../src/jsonUtils'
-import link, { LinkError } from '../src/linker'
+import link from '../src/linker'
+import { Body, Class, Closure, Describe, Environment, Field, Import, Method, Mixin, NamedArgument, Package, Parameter, ParameterizedType, Reference, Return, Singleton, Test, Variable } from '../src/model'
 import wre from '../src/wre/wre.json'
 import { linkerAssertions } from './assertions'
 
@@ -79,7 +79,6 @@ describe('Wollok linker', () => {
         new Package({
           name: 'A',
           members: [
-            new Class({ name: 'X' }),
             new Class({ name: 'Y' }),
           ],
         }),
@@ -92,9 +91,8 @@ describe('Wollok linker', () => {
       ])
     })
 
-    it('should recursively merge same name packages into a single package', () => {
+    it('should recursively override same name packages with new package', () => {
       [
-        ...WRE.members,
         new Package({
           name: 'A',
           members: [
@@ -104,6 +102,7 @@ describe('Wollok linker', () => {
                 new Class({ name: 'X' }),
               ],
             }),
+            new Package({ name: 'C' }),
           ],
         }),
         new Package({
@@ -117,20 +116,21 @@ describe('Wollok linker', () => {
             }),
           ],
         }),
+        ...WRE.members,
       ].should.be.linkedInto([
-        ...WRE.members,
         new Package({
           name: 'A',
           members: [
+            new Package({ name: 'C' }),
             new Package({
               name: 'B',
               members: [
-                new Class({ name: 'X' }),
                 new Class({ name: 'Y' }),
               ],
             }),
           ],
         }),
+        ...WRE.members,
       ])
     })
 
@@ -176,15 +176,15 @@ describe('Wollok linker', () => {
       ], baseEnvironment)
 
       const p = nextEnvironment.members[1]
-      const Y = p.members[1]
+      const Y = p.members[0]
 
+      p.members.should.have.lengthOf(1)
       nextEnvironment.getNodeByFQN('p').should.equal(p)
       nextEnvironment.getNodeByFQN('p.Y').should.equal(Y)
     })
 
-    it('should merge package imports', () => {
+    it('should replace merged packages imports', () => {
       [
-        ...WRE.members,
         new Package({
           name: 'A',
           imports: [
@@ -203,12 +203,10 @@ describe('Wollok linker', () => {
         new Package({ name: 'C' }),
         new Package({ name: 'D' }),
       ].should.be.linkedInto([
-        ...WRE.members,
         new Package({
           name: 'A',
           imports: [
             new Import({ isGeneric: true, entity: new Reference({ name: 'B' }) }),
-            new Import({ isGeneric: true, entity: new Reference({ name: 'C' }) }),
             new Import({ isGeneric: true, entity: new Reference({ name: 'D' }) }),
           ],
         }),
@@ -218,7 +216,20 @@ describe('Wollok linker', () => {
       ])
     })
 
+    it('should replace merged packages problems', () => {
+      [
+        new Package({
+          name: 'A',
+          problems: [{ code: 'ERROR', level: 'error', values: [] }],
+        }),
+        new Package({ name: 'A' }),
+      ].should.be.linkedInto([
+        new Package({ name: 'A' }),
+      ])
+    })
+
   })
+
 
   it('should assign an id to all nodes', () => {
     const environment = link([
@@ -247,7 +258,7 @@ describe('Wollok linker', () => {
           members: [
             new Class({
               name: 'C',
-              supertypes:[new ParameterizedType({ reference: new Reference({ name: 'wollok.lang.Object' }) })],
+              supertypes: [new ParameterizedType({ reference: new Reference({ name: 'wollok.lang.Object' }) })],
               members: [
                 new Field({ name: 'f', isConstant: true, value: new Reference({ name: 'C' }) }),
                 new Field({ name: 'g', isConstant: true, value: new Reference({ name: 'p' }) }),
@@ -663,18 +674,16 @@ describe('Wollok linker', () => {
   describe('error handling', () => {
 
     it('should recover from missing reference in imports', () => {
-      const environment = link([
+      link([
         new Package({
           name: 'p',
           imports: [new Import({ entity: new Reference({ name: 'q.A' }) })],
         }),
       ], WRE)
-
-      environment.getNodeByFQN<Package>('p').imports[0].entity.problems!.should.deep.equal([new LinkError('missingReference')])
     })
 
     it('should recover from missing reference in superclass', () => {
-      const environment = link([
+      link([
         new Package({
           name: 'p',
           members: [
@@ -682,21 +691,17 @@ describe('Wollok linker', () => {
           ],
         }),
       ], WRE)
-
-      environment.getNodeByFQN<Class>('p.C').supertypes[0].reference.problems!.should.deep.equal([new LinkError('missingReference')])
     })
 
     it('should recover from missing reference in mixin', () => {
-      const environment = link([
+      link([
         new Package({
           name: 'p',
           members: [
-            new Class({ name: 'C', supertypes: [new ParameterizedType({ reference: new Reference({ name: 'M' }) })] }),
+            new Mixin({ name: 'M1', supertypes: [new ParameterizedType({ reference: new Reference({ name: 'M2' }) })] }),
           ],
         }),
       ], WRE)
-
-      environment.getNodeByFQN<Class>('p.C').supertypes[0].reference.problems!.should.deep.equal([new LinkError('missingReference')])
     })
 
     it('should not crash if a class inherits from itself', () => {